requirements.txt
tensorflow==2.12.1
keras==2.12.0
matplotlib==3.7.2
numpy==1.24.3
pandas==2.0.3
PyYAML==6.0.1
scikit_learn==1.3.0
seaborn==0.12.2
joblib~=1.4.2
scikit-learn~=1.3.0